% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_pandoc.R
\name{as_Pandoc}
\alias{as_Pandoc}
\title{Convert a Pandoc Element object to a full Pandoc object}
\usage{
as_Pandoc(x)
}
\arguments{
\item{x}{a Pandoc Element object.}
}
\description{
Convert a Pandoc Element object to a full Pandoc object
}
