% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{ast_filter}
\alias{ast_filter}
\title{Apply filters to a Pandoc object}
\usage{
ast_filter(x, ...)
}
\arguments{
\item{x}{a Pandoc object.}

\item{...}{Named filter functions, or (unnamed) lists of named filter
functions.}
}
\description{
Apply filters to a Pandoc object
}
\examples{

# Read in some Markdown content from a string
ast <- md_ast(text = "
## This is a section

Some *text* in a *paragraph*.

### A subsection

* A bullet item
* Another bullet item
")

# Print AST
ast

# Apply filters
ast <- ast_filter(ast,
  # Convert strings to upper case
  Str = function(x) {
    x$c <- tolower(x$c)
    x
  },
  # Convert italic text to bold
  Emph = function(x) {
    Strong(x$c)
  }
)

# Print AST again

# Convert to Markdown (printed at console)
ast_md(ast)

}
\seealso{
\code{\link[=script_filter]{script_filter()}} for use in a Pandoc script.
}
