% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ast.R
\name{ast_md}
\alias{ast_md}
\title{Given a Pandoc AST, write to a file}
\usage{
ast_md(x, outfile = NULL)
}
\arguments{
\item{x}{AST object.}

\item{outfile}{Output filename.}
}
\description{
Given a Pandoc AST, write to a file
}
