% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_utils.R
\name{el_get_class}
\alias{el_get_class}
\alias{el_get_keyvals}
\title{Extract information from the Attr of an Element}
\usage{
el_get_class(x)

el_get_keyvals(x)
}
\arguments{
\item{x}{A Block or Inline Element object which has an Attr child.}
}
\description{
Extract information from the Attr of an Element
}
\examples{
h <- Header(
  attr = Attr(
    classes = Texts("x", "y"),
    attributes = TextText_s(list("a", "a_value"), list("b", "b_value"))
  )
)

el_get_class(h)

el_get_keyvals(h)

}
