% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ast.R
\name{json_ast}
\alias{json_ast}
\title{Read in Pandoc AST JSON and return the Pandoc AST as an R object}
\usage{
json_ast(file = NULL, text = NULL, classify = TRUE)
}
\arguments{
\item{file}{A filename. Cannot be used with \code{text}.}

\item{text}{Pandoc AST JSON text, or a connection object. Cannot be used with
\code{file}.}

\item{classify}{Add classes}
}
\description{
Read in Pandoc AST JSON and return the Pandoc AST as an R object
}
\seealso{
md_ast
}
