% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ast.R
\name{md_ast}
\alias{md_ast}
\title{Read in Markdown content and return the Pandoc AST}
\usage{
md_ast(file = NULL, text = NULL, collapse_strings = TRUE, classify = TRUE)
}
\arguments{
\item{file}{A filename. Cannot be used with \code{text}.}

\item{text}{Markdown text. Cannot be used with \code{file}.}
}
\description{
Read in Markdown content and return the Pandoc AST
}
