% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_tree.R
\name{mixed_to_rawblocks}
\alias{mixed_to_rawblocks}
\title{Convert an HTML tag object with Pandoc elements to a Pandoc Blocks object.}
\usage{
mixed_to_rawblocks(x)
}
\arguments{
\item{x}{An html tag obect that also contains Pandoc Element objects.}
}
\description{
Convert an HTML tag object with Pandoc elements to a Pandoc Blocks object.
}
