% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarker-package.R
\docType{package}
\name{remarker-package}
\alias{remarker}
\alias{remarker-package}
\title{remarker: Translate Markdown to R data structures.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Winston Chang \email{winston@rstudio.com}

}
\keyword{internal}
