% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{script_filter}
\alias{script_filter}
\title{Apply filters in a script}
\usage{
script_filter(..., input = file("stdin", open = "rb"), output = stdout())
}
\arguments{
\item{input}{A connection or filename. Note that the default, \code{file("stdin", open="rb")}, differs from \code{stdin()}. The former reads from the C-level
stdin, so it will read data piped to the R process; the latter reads from
the console input, and may not read from data piped to the process.}

\item{output}{Output connection or filename. Defaults to \code{stdout()}.}
}
\description{
Read Pandoc AST (as JSON), apply filters, and then write the output.
Normally, this function is used in scripts, and the input and output are
stdin and stdout.
}
\details{
Below is an example script\preformatted{library(remarker)

script_filter(
  Str = function(x) \{
    if (x$c == "\{\{helloworld_r\}\}") \{
      Emph(Inlines(Str("Hello, world (R)")))
    \}
  \}
)
}
}
